requirements.txt
fastapi==0.95.0
fastapi-cache2==0.2.1
httpx==0.23.3
requests==2.28.2
uvicorn==0.21.1
